- Jorrit merged 25801:25900 from trunk
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/soc/imposters@26320 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csgeom/vector3.h
@@ -314,7 +314,7 @@ public:
   }
 
   /// Return index of the largest axis
-  inline size_t DominantAxis () const
+  inline int DominantAxis () const
   {
     if (fabsf (x) > fabsf (y) && fabsf (x) > fabsf (z))
       return CS_AXIS_X;
